requirements.txt
# Core ML and Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
imbalanced-learn>=0.11.0

# Feature Extraction - Audio
librosa>=0.10.0
python-speech-features>=0.6
praat-parselmouth>=0.4.3
soundfile>=0.12.0

# Feature Extraction - Image/Video
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0
gunicorn>=21.2.0

# Data Storage and Configuration
pyyaml>=6.0
joblib>=1.3.0

# Jupyter and Notebook Support
jupyter>=1.0.0
ipykernel>=6.23.0
notebook>=6.5.4

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

